64-bit version of Orca

+++ Source/Cocoa Extensions/NSSplitView+Extensions.m
@@ -53,10 +53,10 @@
         NSString*               key = [self ccd__keyForLayoutName: name];
         NSMutableArray* viewRects = [[NSUserDefaults standardUserDefaults] objectForKey: key];
         NSArray*                views = [self subviews];
-        int                             i, count;
+        uint32_t           i, count;
         NSRect                  frame;
 
-        count = MIN( [viewRects count], [views count] );
+        count = (uint32_t)(MIN( [viewRects count], [views count] ));
 
         for( i = 0; i < count; i++ )
         {
