+++ Source/Objects/Data Processing/Filter/FilterScript.h
@@ -20,7 +20,7 @@
 #ifndef FilterScript_h
 #define  FilterScript_h
 
-extern int FilterScriptYYINPUT(char* theBuffer,int maxSize);
+extern int FilterScriptYYINPUT(char* theBuffer,unsigned long maxSize);
 #undef YY_INPUT
 #define YY_INPUT(b,r,s) (r = FilterScriptYYINPUT(b,s))
 
@@ -44,7 +44,7 @@ enum {
 
 /* constants */
 typedef struct {
-    unsigned long value;                  /* value of constant */
+    uint32_t value;                  /* value of constant */
 } conNodeType;
 
 /* identifiers */
