+++ Source/Objects/Hardware/IPE/IpeV4 SLT/PMC_Link.h
@@ -25,13 +25,13 @@
 }
 
 #pragma mark •••Accessors
-- (void) readLongBlockPmc:(unsigned long *) buffer
-			 atAddress:(unsigned int) aPmcAddress
-			 numToRead:(unsigned int) numberLongs;
+- (void) readLongBlockPmc:(uint32_t *) buffer
+			 atAddress:(uint32_t) aPmcAddress
+			 numToRead:(uint32_t) numberLongs;
 
-- (void) writeLongBlockPmc:(unsigned long*) buffer
-			 atAddress:(unsigned int) aPmcAddress
-			 numToWrite:(unsigned int)  numberLongs;
+- (void) writeLongBlockPmc:(uint32_t*) buffer
+			 atAddress:(uint32_t) aPmcAddress
+			 numToWrite:(uint32_t)  numberLongs;
 @end
