+++ Source/Objects/Hardware/Serial/VXM/ORVXMModel.h
@@ -34,7 +34,7 @@
 @private
     NSMutableString* buffer;
     NSMutableArray*	motors;
-    unsigned long	dataId;
+    uint32_t	dataId;
     NSMutableArray* motorToQueryStack;
     NSMutableArray* cmdList;
     BOOL			displayRaw;
@@ -93,15 +93,15 @@
 - (void) setDisplayRaw:(BOOL)aDisplayRaw;
 - (NSArray*) motors;
 - (ORVXMMotor*) motor:(int)aMotor;
-- (void) addItem:(id)anItem atIndex:(int)anIndex;
-- (void) removeItemAtIndex:(int) anIndex;
+- (void) addItem:(id)anItem atIndex:(NSInteger)anIndex;
+- (void) removeItemAtIndex:(NSInteger) anIndex;
 - (BOOL) isMoving;
 
 #pragma mark ***Data Records
 - (void) appendDataDescription:(ORDataPacket*)aDataPacket userInfo:(NSDictionary*)userInfo;
 - (NSDictionary*) dataRecordDescription;
-- (unsigned long) dataId;
-- (void) setDataId: (unsigned long) DataId;
+- (uint32_t) dataId;
+- (void) setDataId: (uint32_t) DataId;
 - (void) setDataIds:(id)assigner;
 - (void) syncDataIdsWith:(id)anotherVXM;
 - (void) shipMotorState:(id)aMotor;
