+++ Source/Objects/Hardware/Vme/Vme Base Classes/ORVmeBusProtocol.h
@@ -27,47 +27,47 @@
 - (void) resetContrl;
 - (void) checkStatusErrors;
 
-- (void) readLongBlock:(unsigned long *) readAddress
-			 atAddress:(unsigned int) vmeAddress
-			 numToRead:(unsigned int) numberLongs
+- (void) readLongBlock:(uint32_t *) readAddress
+			 atAddress:(uint32_t) vmeAddress
+			 numToRead:(uint32_t) numberLongs
 			withAddMod:(unsigned short) addressModifier
 		 usingAddSpace:(unsigned short) addressSpace;
 
-- (void) readLong:(unsigned long *) readAddress
-		atAddress:(unsigned long) vmeAddress
-	  timesToRead:(unsigned int) numberLongs
+- (void) readLong:(uint32_t *) readAddress
+		atAddress:(uint32_t) vmeAddress
+	  timesToRead:(uint32_t) numberLongs
 	   withAddMod:(unsigned short) addModifier
 	usingAddSpace:(unsigned short) addressSpace;
 
 
-- (void) writeLongBlock:(unsigned long *) writeAddress
-			  atAddress:(unsigned int) vmeAddress
-			 numToWrite:(unsigned int) numberLongs
+- (void) writeLongBlock:(uint32_t *) writeAddress
+			  atAddress:(uint32_t) vmeAddress
+			 numToWrite:(uint32_t) numberLongs
 			 withAddMod:(unsigned short) addressModifier
 		  usingAddSpace:(unsigned short) addressSpace;
 
 - (void) readByteBlock:(unsigned char *) readAddress
-			 atAddress:(unsigned int) vmeAddress
-			 numToRead:(unsigned int) numberBytes
+			 atAddress:(uint32_t) vmeAddress
+			 numToRead:(uint32_t) numberBytes
 			withAddMod:(unsigned short) addressModifier
 		 usingAddSpace:(unsigned short) addressSpace;
 
 - (void) writeByteBlock:(unsigned char *) writeAddress
-			  atAddress:(unsigned int) vmeAddress
-			 numToWrite:(unsigned int) numberBytes
+			  atAddress:(uint32_t) vmeAddress
+			 numToWrite:(uint32_t) numberBytes
 			 withAddMod:(unsigned short) addressModifier
 		  usingAddSpace:(unsigned short) addressSpace;
 
 
 - (void) readWordBlock:(unsigned short *) readAddress
-			 atAddress:(unsigned int) vmeAddress
-			 numToRead:(unsigned int) numberWords
+			 atAddress:(uint32_t) vmeAddress
+			 numToRead:(uint32_t) numberWords
 			withAddMod:(unsigned short) addressModifier
 		 usingAddSpace:(unsigned short) addressSpace;
 
 - (void) writeWordBlock:(unsigned short *) writeAddress
-			  atAddress:(unsigned int) vmeAddress
-			 numToWrite:(unsigned int) numberWords
+			  atAddress:(uint32_t) vmeAddress
+			 numToWrite:(uint32_t) numberWords
 			 withAddMod:(unsigned short) addressModifier
 		  usingAddSpace:(unsigned short) addressSpace;
